+++ class/actions_massaction.class.php
@@ -150,6 +150,7 @@ class Actionsmassaction extends \massaction\RetroCompatCommonHookActions
 
 		$this->resprints = $toprint;
 	}
+
 	/**
 	 * Overloading the doMassActions function : replacing the parent's function with the one below
 	 *
@@ -168,6 +169,7 @@ class Actionsmassaction extends \massaction\RetroCompatCommonHookActions
 
 		$TContext = explode(":", $parameters['context']);
 		$confirm = GETPOST('confirm', 'alphanohtml');
+		$massActionToken = GETPOST('massaction_token', 'alphanohtml');
 
 		$error = 0; // Error counter
 		$errormsg = ''; // Error message
@@ -389,6 +391,7 @@ class Actionsmassaction extends \massaction\RetroCompatCommonHookActions
 			|| in_array('contactlist', $TContext)
 			|| in_array('memberlist', $TContext)
 			|| in_array('userlist', $TContext)) {
+
 			if (isModEnabled('mailing') && $user->hasRight('mailing', 'creer')) {
 				$label = '<span class="fa fa-envelope-o paddingrightonly"></span> ' . $langs->trans("MassActionLinktoMailing");
 				$this->resprints .= '<option value="linktomailing" data-html="' . dol_escape_htmltag($label) . '">' . $label . '</option>';
@@ -439,6 +442,8 @@ class Actionsmassaction extends \massaction\RetroCompatCommonHookActions
 
 		return $select;
 	}
+
+
 	/**
 	 * Processes mass actions on document lines.
 	 *
@@ -462,6 +467,9 @@ class Actionsmassaction extends \massaction\RetroCompatCommonHookActions
 
 		if (!empty($commonContexts)) {
 			require_once __DIR__ . '/massaction.class.php';
+			if (empty($massActionToken)) {
+				$massActionToken = MassAction::getMassActionToken();
+			}
 
 			$langs->load('massaction@massaction');
 			$massAction = new MassAction($this->db, $object);
@@ -470,6 +478,10 @@ class Actionsmassaction extends \massaction\RetroCompatCommonHookActions
 			$TSelectedLines = explode(',', $selectedLines);
 
 			$confirm = GETPOST('confirm', 'alpha');
+
+			if ($action == 'preSelectSupplierPrice' && !$confirm && !GETPOST('sendit', 'alpha') && !GETPOST('remove_massaction_file', 'alpha')) {
+				MassAction::cleanupTemporaryUploads(array(), '', $massActionToken);
+			}
 
 			if ($action == 'delete_lines' && $confirm == 'yes') {
 				$TRowIds = array_column($object->lines, 'rowid');
@@ -490,6 +502,7 @@ class Actionsmassaction extends \massaction\RetroCompatCommonHookActions
 				$massAction->handleErrors($TSelectedLines, $massAction->TErrors, $action);
 
 				$action = '';
+
 			}
 
 			if (($action == 'edit_quantity' || $action == 'edit_margin') && $confirm == 'yes') {
@@ -513,6 +526,7 @@ class Actionsmassaction extends \massaction\RetroCompatCommonHookActions
 						$index = array_search(intval($selectedLine), $TRowIds);
 
 						$massAction->updateLine($index, $quantity, $marge_tx);
+
 					}
 				}
 
@@ -525,16 +539,48 @@ class Actionsmassaction extends \massaction\RetroCompatCommonHookActions
 				$massAction->handleErrors($TSelectedLines, $errors, $action);
 
 				$action = '';
+
 			}
 			// Supplier price request management
+			if ($action == 'preSelectSupplierPrice' && GETPOST('sendit', 'alpha')) {
+				$uploadResult = MassAction::persistUploadedFiles($_FILES['massaction_files'] ?? array(), $massActionToken);
+				if (!empty($uploadResult['errors'])) {
+					setEventMessages($langs->trans('ErrorFileNotUploaded'), $uploadResult['errors'], 'errors');
+				} elseif (!empty($uploadResult['files'])) {
+					setEventMessage($langs->trans('FileTransmitted'));
+				}
+				$action = 'preSelectSupplierPrice';
+			} elseif ($action == 'preSelectSupplierPrice' && GETPOST('remove_massaction_file', 'alpha')) {
+				$fileToRemove = GETPOST('remove_massaction_file', 'alpha');
+				$resultRemove = MassAction::removePersistedUpload($fileToRemove, $massActionToken);
+				if ($resultRemove < 0) {
+					setEventMessage($langs->trans('ErrorFailedToDeleteFile', $fileToRemove), 'errors');
+				} else {
+					setEventMessage($langs->trans('FileDeleted'));
+				}
+				$action = 'preSelectSupplierPrice';
+			}
 			if ($action == 'createSupplierPrice') {
+				if ($confirm !== 'yes') {
+					MassAction::cleanupTemporaryUploads(array(), '', $massActionToken);
+					$action = '';
+					return 0;
+				}
+
 				$supplierIds = GETPOST('supplierid', 'array');
 				$templateId = GETPOST('model_mail', 'int');
-				$massAction->handleCreateSupplierPriceAction($object, $TSelectedLines, $supplierIds, (int) $templateId);
+				$deliveryDate = null;
+				if (GETPOSTINT('maxresponse_year')) {
+					$deliveryDate = dol_mktime(12, 0, 0, GETPOSTINT('maxresponse_month'), GETPOSTINT('maxresponse_day'), GETPOSTINT('maxresponse_year'));
+				}
+				// Attachments are already persisted during preSelectSupplierPrice
+				$massAction->handleCreateSupplierPriceAction($object, $TSelectedLines, $supplierIds, (int) $templateId, $deliveryDate, array(), $massActionToken);
 			}
 		}
 		return 0;
 	}
+
+
 	/**
 	 * Injects checkboxes and mass action buttons into document line tables.
 	 *
@@ -652,11 +698,10 @@ class Actionsmassaction extends \massaction\RetroCompatCommonHookActions
 					var TSelectedLines = selectedLines.split(',');
 
 					var currentAction = "<?php echo $action ?>";
-
 					var toShow = formConfirm !== '' ? formConfirm : (massActionButton || '');
 
 					var form = `
-						<form method="post" id="searchFormList" action="` + action + `">
+						<form method="post" id="massactionForm" action="` + action + `">
 							<input type="hidden" name="token" value="` + token + `">
 							<input type="hidden" name="selectedLines" id="selectedLines" value="` + selectedLines + `">
 							<input type="hidden" name="action" value="">
@@ -669,6 +714,29 @@ class Actionsmassaction extends \massaction\RetroCompatCommonHookActions
 					$('#addproduct:last-child').before(form);
 
 					showCheckboxes();
+
+					if (currentAction === 'preSelectSupplierPrice') {
+						$('#massactionForm').attr('enctype', 'multipart/form-data');
+						$('.massaction-remove-file').on('click', function (e) {
+							e.preventDefault();
+							var fname = $(this).data('filename');
+							$('.removedfilehidden').val(fname);
+							$('input[name="action"]').val('preSelectSupplierPrice');
+							$('#confirm').val('no');
+							$('#massactionForm').submit();
+						});
+						$('.massaction-file-input').on('change', function () {
+							if ($(this).val()) {
+								$('input[name="action"]').val('preSelectSupplierPrice');
+								$('#confirm').val('no');
+								$('input[name="sendit"]').val('1');
+								$('#massactionForm').submit();
+							}
+						});
+						$('.confirmvalidatebutton').on('click', function () {
+							$('input[name="action"]').val('createSupplierPrice');
+						});
+					}
 
 					// Cocher automatiquement les cases à cocher si l'action est predelete ou edit_margin ou edit_quantity
 					if (currentAction === 'predelete' || currentAction === 'preeditquantity' || currentAction === 'preeditmargin' || currentAction === 'preSelectSupplierPrice') {
@@ -711,6 +779,7 @@ class Actionsmassaction extends \massaction\RetroCompatCommonHookActions
 			</script>
 
 			<?php
+
 		}
 
 		return 0;
@@ -783,7 +852,7 @@ class Actionsmassaction extends \massaction\RetroCompatCommonHookActions
 							$('#pop-split').remove();
 							$('body').append('<div id="pop-split"></div>');
 
-							$.get('<?php echo dol_buildpath('/massaction/script/showLines.php', 1).'?id='.$object->id.'&element='.$object->element.'&selectedLines='.$selectedLines ?>', function(data) {
+							$.get('<?php echo dol_buildpath('/massaction/script/showLines.php',1).'?id='.$object->id.'&element='.$object->element.'&selectedLines='.$selectedLines ?>', function(data) {
 								$('#pop-split').html(data)
 
 								$('#pop-split').dialog({
@@ -870,4 +939,6 @@ class Actionsmassaction extends \massaction\RetroCompatCommonHookActions
 		}
 		return 0;
 	}
+
+
 }