+++ core/modules/modmassaction.class.php
@@ -61,7 +61,7 @@ class modmassaction extends DolibarrModules
 		// Module description, used if translation string 'ModuleXXXDesc' not found (where XXX is value of numeric property 'numero' of module)
 		$this->description = "ModuleMassActionDesc";
 		// Possible values for version are: 'development', 'experimental', 'dolibarr' or version
-		$this->version = '1.7.1';
+		$this->version = '1.7.5';
 		// Key used in llx_const table to save module status enabled/disabled (where MYMODULE is value of property name of module in uppercase)
 		$this->const_name = 'MAIN_MODULE_'.strtoupper($this->name);
 		// Where to store the module in setup page (0=common,1=interface,2=others,3=very specific)
@@ -121,6 +121,7 @@ class modmassaction extends DolibarrModules
 		// Main menu entries
 		$this->menu = array();			// List of menus to add
 		$r=0;
+
 	}
 
 	/**
@@ -135,7 +136,7 @@ class modmassaction extends DolibarrModules
 	{
 		$sql = array();
 
-		define('INC_FROM_DOLIBARR', true);
+		define('INC_FROM_DOLIBARR',true);
 
 		$this->addEmailTemplateForSupplierProposal();
 
@@ -175,6 +176,7 @@ class modmassaction extends DolibarrModules
 		$obj = $this->db->fetch_object($res);
 
 		if ($obj->total == 0) {
+
 			$topic = $langs->transnoentities('MassActionSupplierPriceRequest') . ' __REF__';
 			$content = $langs->transnoentities('MassActionHello') . " __THIRDPARTY_NAME__,\n\n";
 			$content .= $langs->transnoentities('MassActionPleaseFindAttachedOurPriceRequest') . " __REF__.\n\n";
@@ -193,4 +195,6 @@ class modmassaction extends DolibarrModules
 			$this->db->query($sql);
 		}
 	}
+
+
 }